src/requirements.txt
uvicorn
fastapi
redis
celery
torch
pydantic
python-dotenv
numpy
python-multipart